First version of power simulation ShinyApp
This is the first working version of the power simulation app. It only allows varying alpha level, maximum datasetsize, and number of repetitions for the power simulation. It uses two randomly sampled normal distributions to compare. The goal is to extend this to more applicable statistical tests and datadistributions seen in ML for medical imaging segmentation.